+++ drizzle/0005_large_nicolaos.sql
@@ -1,0 +1,5 @@
+CREATE TABLE "song-download-table" (
+	"id" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name "song-download-table_id_seq" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),
+	"songcid" varchar(255) NOT NULL,
+	"downloadPath" text NOT NULL
+);
