+++ drizzle/meta/_journal.json
@@ -36,6 +36,13 @@
       "when": 1760319870598,
       "tag": "0004_burly_outlaw_kid",
       "breakpoints": true
+    },
+    {
+      "idx": 5,
+      "version": "7",
+      "when": 1767284179533,
+      "tag": "0005_large_nicolaos",
+      "breakpoints": true
     }
   ]
 }
