+++ migrations.json
@@ -1,1 +1,1 @@
-[{"sql":["CREATE TABLE \"playback-count\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"playback-count_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"count\" integer NOT NULL\n);\n"],"bps":true,"folderMillis":1746214727681,"hash":"30aa92e016389862d078c10dae9578bf11734556f21bf3828db7da9ed65184a8"},{"sql":["CREATE TABLE \"temp-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"temp-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL\n);\n","\nALTER TABLE \"playback-count\" ADD COLUMN \"lastPlayed\" integer;"],"bps":true,"folderMillis":1747071747438,"hash":"17e69f61a2f2e98c46765e372d2fe3faebe890d6182b458ad1ab5d525eae416f"},{"sql":["CREATE TABLE \"abrepeat-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"abrepeat-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"a\" real,\n\t\"b\" real,\n\tCONSTRAINT \"abrepeat-table_songcid_unique\" UNIQUE(\"songcid\")\n);\n","\nCREATE TABLE \"lyric-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"lyric-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songId\" varchar(255) NOT NULL,\n\t\"lyricKey\" varchar(255) NOT NULL,\n\t\"lyricOffset\" real NOT NULL,\n\t\"lyric\" text NOT NULL,\n\t\"source\" varchar(255),\n\tCONSTRAINT \"lyric-table_songId_unique\" UNIQUE(\"songId\")\n);\n","\nCREATE TABLE \"r128gain-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"r128gain-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"r128gain\" real,\n\tCONSTRAINT \"r128gain-table_songcid_unique\" UNIQUE(\"songcid\")\n);\n"],"bps":true,"folderMillis":1747168543082,"hash":"d986a047d64c11aca3ca5c5617ac6b5861401086744825bcab013660d989a6db"},{"sql":["CREATE TABLE \"song-beat-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"song-beat-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"beat\" text NOT NULL\n);\n"],"bps":true,"folderMillis":1757528305658,"hash":"31eeeb11a08214f504a9df6ec83c8203e27639bdb7381cf12b9024067b4abe86"},{"sql":["ALTER TABLE \"playback-count\" ALTER COLUMN \"lastPlayed\" SET DATA TYPE bigint;"],"bps":true,"folderMillis":1760319870598,"hash":"0abed23c1545cfdb9b68477b486da5ab2a7df5d5a8a8c81f0a7914d92a491a3c"}]
+[{"sql":["CREATE TABLE \"playback-count\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"playback-count_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"count\" integer NOT NULL\n);\n"],"bps":true,"folderMillis":1746214727681,"hash":"30aa92e016389862d078c10dae9578bf11734556f21bf3828db7da9ed65184a8"},{"sql":["CREATE TABLE \"temp-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"temp-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL\n);\n","\nALTER TABLE \"playback-count\" ADD COLUMN \"lastPlayed\" integer;"],"bps":true,"folderMillis":1747071747438,"hash":"17e69f61a2f2e98c46765e372d2fe3faebe890d6182b458ad1ab5d525eae416f"},{"sql":["CREATE TABLE \"abrepeat-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"abrepeat-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"a\" real,\n\t\"b\" real,\n\tCONSTRAINT \"abrepeat-table_songcid_unique\" UNIQUE(\"songcid\")\n);\n","\nCREATE TABLE \"lyric-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"lyric-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songId\" varchar(255) NOT NULL,\n\t\"lyricKey\" varchar(255) NOT NULL,\n\t\"lyricOffset\" real NOT NULL,\n\t\"lyric\" text NOT NULL,\n\t\"source\" varchar(255),\n\tCONSTRAINT \"lyric-table_songId_unique\" UNIQUE(\"songId\")\n);\n","\nCREATE TABLE \"r128gain-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"r128gain-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"r128gain\" real,\n\tCONSTRAINT \"r128gain-table_songcid_unique\" UNIQUE(\"songcid\")\n);\n"],"bps":true,"folderMillis":1747168543082,"hash":"d986a047d64c11aca3ca5c5617ac6b5861401086744825bcab013660d989a6db"},{"sql":["CREATE TABLE \"song-beat-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"song-beat-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"beat\" text NOT NULL\n);\n"],"bps":true,"folderMillis":1757528305658,"hash":"31eeeb11a08214f504a9df6ec83c8203e27639bdb7381cf12b9024067b4abe86"},{"sql":["ALTER TABLE \"playback-count\" ALTER COLUMN \"lastPlayed\" SET DATA TYPE bigint;"],"bps":true,"folderMillis":1760319870598,"hash":"0abed23c1545cfdb9b68477b486da5ab2a7df5d5a8a8c81f0a7914d92a491a3c"},{"sql":["CREATE TABLE \"song-download-table\" (\n\t\"id\" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name \"song-download-table_id_seq\" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),\n\t\"songcid\" varchar(255) NOT NULL,\n\t\"downloadPath\" text NOT NULL\n);\n"],"bps":true,"folderMillis":1767284179533,"hash":"6ea3cd88d3ef41c2f2eff2088eae07d1d6724943dd85b8b08ebcffd223744093"}]
